pyproject.toml
[project]
name = "scrap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "httpx>=0.28.1",
    "lxml>=5.4.0",
    "mcp[cli]>=1.9.2",
    "pytest-playwright>=0.7.0",
    "requests>=2.32.3",
    "tf-playwright-stealth>=1.1.2",
]
